💀

+++ app/src/main/java/com/cocktailbar/presentation/cocktails/CocktailEditState.kt
@@ -2,7 +2,7 @@ package com.cocktailbar.presentation.cocktails
 
 data class CocktailEditState(
     val image: String? = null,
-    val imageLoaderProgress: Float = LOADER_PROGRESS_COMPLETED,
+    val imageLoaderProgress: Int = LOADER_PROGRESS_COMPLETED,
     val title: String = "",
     val description: String = "",
     val recipe: String = "",
@@ -10,6 +10,6 @@ data class CocktailEditState(
     val saveLoading: Boolean = false,
 ) {
     companion object {
-        const val LOADER_PROGRESS_COMPLETED = 100f
+        const val LOADER_PROGRESS_COMPLETED = 100
     }
 }
